Refactor testing to merge checkUtil.js with util.js and replace asssetTransfer with transactionExecute
